fix: idempotent release and attempt to fix tests on slower HW specs

+++ evita_store/evita_store_key_value/src/main/java/io/evitadb/store/offsetIndex/io/WriteOnlyOffHeapWithFileBackupHandle.java
@@ -247,9 +247,11 @@ public class WriteOnlyOffHeapWithFileBackupHandle implements WriteOnlyHandle {
 	 * This method closes the file output stream, frees the observable output for the target file, and deletes the temporary file.
 	 */
 	private void releaseTemporaryFile() {
-		this.fileOutput.close();
-		this.observableOutputKeeper.close(this.targetFile);
-		Assert.isPremiseValid(getTargetFile(this.targetFile).delete(), "Failed to delete temporary file `" + this.targetFile + "`!");
+		if (this.fileOutput != null) {
+			this.fileOutput.close();
+			this.observableOutputKeeper.close(this.targetFile);
+			Assert.isPremiseValid(getTargetFile(this.targetFile).delete(), "Failed to delete temporary file `" + this.targetFile + "`!");
+		}
 		this.fileOutput = null;
 	}
